Fixes for reweighting, extra RegularizedResampler function, more output, BoldModel Bugfix, removed warnings.
When a particle gets weighted to 0, it now gets a "default" value that
won't go off the deep end.

Added a function that will accept a covariance to bet given when
resampling rather than having it be calculated by the resampling
algorithm itself. This could permit pre-stratified resampled values for
the covariance to be used in cases of extreme particle impoverishment.

Boldmodel had a problem where arguments were out of order. Misc warnings

+++ code/include/tools.h
@@ -185,8 +185,8 @@ typename itk::OrientedImage< T, 4 >::Pointer prune(
     typedef itk::OrientedImage< T, 4>  ImageType;
     
     /* Check sizes */
-    if(stop >= in->GetRequestedRegion().GetSize()[dir]) 
-        stop = in->GetRequestedRegion().GetSize()[dir]-1;
+    if(stop >= (int)in->GetRequestedRegion().GetSize()[dir]) 
+        stop = (int)in->GetRequestedRegion().GetSize()[dir]-1;
     
     if(start < 0)
         start = 0;
